Cambio de tipo de dato de IdUsuario a long

+++ src/main/java/huellitassoft_web/huellitasoft/entity/User.java
@@ -20,7 +20,7 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_usuario")
-    private Integer idUsuario;
+    private Long idUsuario;
 
     @Column(nullable = false, unique = true)
     private String email;
